Should save messageId in queue (#31)
* Update README.md

* Fixed typo & standardize

* Added nodejs 12 support

* PUBLISH packet should NOT update the outgoing queue
Broker uses `brokerCounter` to track how many packets it emits, as `messageId` will be rotated after it reaches max 65535. If there are two PUBLISH packets which brokerCounter is different but same messageId, the second PUBLISH will update the first slot in the queue, as `outgoingUpdate` is using `for` loop starting at index 0. This will fix the issue. We use `outgoingEnqueue` or `outgoingEnqueueCombi` to enqueue PUBLISH packets, and and use `outgoingUpdate` to update those PUBLSIH items to next state (say PUBREL), so `outgoing` array should not contain PUBLISH, and that's why this issue could be fixed by a check `packet.cmd !== 'publish'`

* Removed console.log

* Should save messageID in queue

* Drop nodejs 6 support

* Use aedes-packet v2.0.0